+++ src/main/java/org/realityforge/bazel/depgen/model/OptionsModel.java
@@ -19,8 +19,8 @@ import org.realityforge.bazel.depgen.util.BazelUtil;
 public final class OptionsModel {
     private static final List<String> SUPPORTED_REPOSITORY_RULE_LOAD_SYMBOLS = List.of("http_file");
 
-    private static final List<String> SUPPORTED_TARGET_RULE_LOAD_SYMBOLS =
-            Arrays.asList("java_binary", "java_import", "java_library", "java_plugin", "java_test", "j2cl_library");
+    private static final List<String> SUPPORTED_TARGET_RULE_LOAD_SYMBOLS = Arrays.asList(
+            "java_binary", "java_import", "java_library", "java_plugin", "java_test", "j2cl_import", "j2cl_library");
 
     private final OptionsConfig _source;